% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplOvCa-package.r
\docType{package}
\name{supplOvCa}
\alias{supplOvCa}
\alias{supplOvCa-package}
\title{supplOvCa.
Annotation for TCGA ovarian cancer patient, extracted from parsing Supplementary files from ovarian cancer articles including PMID 2325762, 21720365}
\description{
supplOvCa.
Annotation for TCGA ovarian cancer patient, extracted from parsing Supplementary files from ovarian cancer articles including PMID 2325762, 21720365
}
